DeleteAuthorCommand updated.

+++ WebApi/Application/AuthorOperations/Commands/DeleteAuthor/DeleteAuthorCommand.cs
@@ -19,6 +19,9 @@ namespace WebApi.Application.AuthorOperations.Commands.DeleteAuthor
             var author = _context.Authors.FirstOrDefault(b => b.Id == AuthorId);
             if (author is null)
                 throw new InvalidOperationException("Yazar mevcut değil!");
+            var bookOfAuthor = _context.Books.Any(b=>b.AuthorId == author.Id);
+            if (!bookOfAuthor)
+                throw new InvalidOperationException("Yazara ait yayında olan kitap var. Önce kitabı silin.");
 
             _context.Authors.Remove(author);
             _context.SaveChanges();
